buildQuestionnaire
