+++ src/selectors/questionnaire/select_questionnaire.ts
@@ -1,8 +1,9 @@
 import { Store } from '../../stores';
 import { selectLocale } from '../locale/select_locale';
-import { denormalizeQuestions, Questionnaire } from '.';
+import { Questionnaire } from '.';
+import { buildQuestionnaire } from './build_questionnaire';
 
 export const selectQuestionnaire = (appStore: Store): Questionnaire => {
     const locale = selectLocale(appStore);
-    return denormalizeQuestions(locale, appStore.questionnaireInStore);
+    return buildQuestionnaire(locale, appStore.questionnaireInStore);
 };
